Update README.md
Python for Data Science (PY0101EN) is an online course offered by IBM.
Completing this course will count towards learning Data Science fundamentals with Python and SQL

In this course I learn: 
(1). What is Python and why is it useful
(2). The application of Python 
(3). How to define variables
(4). Sets and conditional statements in Python
(5). The purpose of having functions in Python
(6). How to operate on files to read and write data in Python
(7). How to use pandas